Put postRequestWithImage function to a file (FE, #459)
Couldn't use useNavigate anymore in the function since it is no longer in a React component. Used window.location.href instead.
